+++ ocpp_wallbox_mqtt/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## [1.0.6.2] - 2026-08-25
+- FIX il merge di ocpp.ini col template copiava le chiavi nuove senza i commenti
+  che le descrivono: restava un elenco di nomi senza spiegazione
 ## [1.0.6.1] - 2026-08-25
 - ADD supporto a 2 wallbox
 ## [1.0.4.14] - 2026-04-20